pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "psw-server-inspector"
version = "0.1.0"
description = "Hardware detection and inventory for server"
readme = "README.md"
requires-python = ">=3.14"
license = "MIT"
authors = [{ name = "Casa Eureka" }]
keywords = ["hardware", "detection", "inventory", "server", "cli"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.14",
    "Topic :: System :: Systems Administration",
]

dependencies = [
    "psutil",
    "pyyaml",
]

[project.scripts]
server-inspector = "psw_server_inspector.main:main"

[project.urls]
Homepage = "https://github.com/casaeureka/psw-server-inspector"
Repository = "https://github.com/casaeureka/psw-server-inspector"
Issues = "https://github.com/casaeureka/psw-server-inspector/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/psw_server_inspector"]

[tool.hatch.build.targets.wheel.sources]
"src" = ""

[tool.hatch.build]
include = [
    "src/psw_server_inspector/**/*.py",
    "src/psw_server_inspector/**/*.sh",
]

[tool.ruff]
target-version = "py314"
line-length = 120

[tool.ruff.lint]
select = ["E", "W", "F", "I", "N", "UP", "B", "C4", "ARG", "SIM"]
ignore = ["E501", "B008"]

[tool.ruff.lint.isort]
section-order = [
    "future",
    "standard-library",
    "third-party",
    "first-party",
    "local-folder",
]
known-first-party = ["psw_server_inspector"]

[tool.mypy]
python_version = "3.14"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
packages = ["psw_server_inspector"]

[dependency-groups]
dev = [
    "mypy",
    "pytest",
    "ruff",
    "types-pyyaml",
]

[[tool.mypy.overrides]]
module = ["colors", "psutil"]
ignore_missing_imports = true
